Fix some warnings
cargo check shows a lot of warnings with this generated code.
This commit fixes some of them.  I also changed the
src/apis/configuration.rs file to take a variable instead of
being hardcoded.  Not sure why it was like that.  There may
be other issues with this library that will show up in testing.