Error when session is expired
The swagger generated code has painted me into a corner where
everything is immutable.  I wanted to have the code just
refresh the session token if it expired but the Rc on
configuration::Configuration prevents that.  So instead this will
now throw an Err(Error::SessionExpired) to prompt the user to
login again.  It's not ideal but I think it'll work for now